share button into own file
