+++ apps/dotcom/client/src/tla/components/TlaEditor/TlaEditorTopRightPanel.tsx
@@ -1,10 +1,9 @@
 import classNames from 'classnames'
-import { forwardRef, useRef } from 'react'
+import { useRef } from 'react'
 import { PeopleMenu, usePassThroughWheelEvents } from 'tldraw'
 import { useCurrentFileId } from '../../hooks/useCurrentFileId'
-import { useTldrawAppUiEvents } from '../../utils/app-ui-events'
-import { F } from '../../utils/i18n'
 import { TlaFileShareMenu } from '../TlaFileShareMenu/TlaFileShareMenu'
+import { TlaShareButton } from '../TlaShareButton/TlaShareButton'
 import { TlaSignUpButton } from './TlaSignUpButton'
 import styles from './top.module.css'
 
@@ -28,30 +27,8 @@ export function TlaEditorTopRightPanel({ isAnonUser }: { isAnonUser: boolean }) 
 		<div ref={ref} className={styles.topRightPanel}>
 			<PeopleMenu displayUserWhenAlone={false} />
 			<TlaFileShareMenu fileId={fileId!} source="file-header">
-				<ShareButton />
+				<TlaShareButton />
 			</TlaFileShareMenu>
 		</div>
 	)
 }
-
-// todo, move styles from globals.css to top.module.css
-
-export const ShareButton = forwardRef<HTMLButtonElement>(function ShareButton(props, ref) {
-	const trackEvent = useTldrawAppUiEvents()
-	return (
-		<button
-			ref={ref}
-			draggable={false}
-			type="button"
-			className="tlui-share-zone__button-wrapper"
-			{...props}
-			onClick={() => trackEvent('open-share-menu', { source: 'file-header' })}
-		>
-			<div className="tlui-button tlui-button__normal tlui-share-zone__button">
-				<span className="tlui-button__label" draggable={false}>
-					<F defaultMessage="Share" />
-				</span>
-			</div>
-		</button>
-	)
-})
